==> 1_face_non/R_analysing/scripts/8_calculating_zscores.Rmd <==
---
title: "calculating_zscores"
author: "Sahar Fatima"
date: "23/06/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

#### note: going to be adapting jenny's steps: https://rpubs.com/jenrichmond/emgz !!
#### note: for faces, 429 = happy, 673 = angry
### load packages 
```{r message=FALSE, warning=FALSE}
library(tidyverse)
library(formattable) # useful for controlling decimal places
library(here)

options(scipen = 999) #set no scientific notation
```

### The process
- first, make data wide using pivot_wider so that brow and cheek are in columns
- use scale() to transform RMS scores into z scores for each participant.
- then make bin wide so that each row represents a single trial and bins 1-10 are in different columns
- make new column for each bin, calculating the difference between that binZ and baselineZ.

# Calculating 
Normally, to center a variable, you would subtract the mean of all data points from each individual data point. With scale(), this can be accomplished in one simple call.

center variable A using the scale() function
`scale(A, center = TRUE, scale = FALSE)`

### 2. Generating Z-Scores

generate z-scores for variable A using the scale() function
`scale(A, center = TRUE, scale = TRUE)`

## read in data 

```{r message=FALSE, warning=FALSE}
faces_data <- read_csv(here("R_analysing", "data", "7_face_false.csv")) %>% # face data 
  select(-outlier1.5SD)

non_data <- read_csv(here("R_analysing", "data", "7_non_false.csv")) %>% # non-face data
  select(-outlier1.5SD)
```

### going to try on one pp first - from faces df

```{r}
pp481 <- faces_data %>%
  filter(pp_no == "pp481")

# make it wide 
widepp481 <- pp481 %>%
  pivot_wider(names_from = muscle, values_from = rms)

# create new variables called zbrow and zcheek using the scale function and specifying center = TRUE and scale = TRUE

widepp481$zbrow = scale(widepp481$brow, center = TRUE, scale = TRUE)
widepp481$zcheek = scale(widepp481$cheek, center = TRUE, scale = TRUE)

# Check that the mean of the new zbrow and zcheek columns is 0 - this is what we want... 
summary(widepp481)

```

it works... going to try on faces and non-face data sets now :) 

## faces
```{r}
# make data wide, use mutate to make new z variables
faces_data_zscore <- faces_data %>%
  pivot_wider(names_from = muscle, 
              values_from = rms) %>%
  group_by(pp_no) %>%
  mutate(Zbrow = scale(brow, center = TRUE, scale = TRUE)) %>%
  mutate(Zcheek = scale(cheek, center = TRUE, scale = TRUE))
  
summary(faces_data_zscore)
```

note: can also make new z-variables manually: 
`data_z_manual <- data_wide %>% 
group_by(pp_no) %>%
mutate(Zbrow = (brow - mean(brow))/sd(brow)) %>%
mutate(Zcheek = (cheek - mean(cheek))/sd(cheek))`

## non-faces 

```{r}
# make data wide, use mutate to make new z variables
non_data_zscore <- non_data %>%
  pivot_wider(names_from = muscle, values_from = rms) %>%
  group_by(pp_no) %>%
  mutate(Zbrow = scale(brow, center = TRUE, scale = TRUE)) %>%
  mutate(Zcheek = scale(cheek, center = TRUE, scale = TRUE))
  
summary(non_data_zscore)
```

# Calculating Difference Scores 
need to calculate difference scores from baseline. Need to make BIN wide to allow for calculations across columns. Its a bit difficult to so that for both muscles at the same time, so lets separate and work out the difference scores for brow and cheek separately.

## facial stimuli 

### faces - brow
```{r}
# add a muscle column to make things easier to join back together later
face_brow_z <- faces_data_zscore %>%
  mutate(muscle = "brow") %>%
  select(pp_no, condition, bin, trial, muscle, Zbrow) 

# Make the bin column wide and rename bin_0 as BL (i.e. baseline)
face_brow_z_wide <- face_brow_z %>%
  pivot_wider(names_from = bin, 
              values_from = Zbrow) %>%
  rename(BL = bin_0)

# mutate() diff scores - Use wide columns to calcuate the difference between each bin column and BL, creating a new set of columns starting with “diff”, drop BL column and all columns starting with bin (i.e. raw z scores).

# note - This df contains for each bin the difference between stimulus and basline, so POSITIVE difference scores = greater activity during STIM than BL and NEGATIVE difference scores = greater activity during BL than STIM

face_brow_z_diff <- face_brow_z_wide %>% 
  mutate(diff_bin1 = bin_1 - BL, 
         diff_bin2 = bin_2 - BL,
         diff_bin3 = bin_3- BL, 
         diff_bin4 = bin_4 - BL,
         diff_bin5 = bin_5 - BL, 
         diff_bin6 = bin_6 - BL,
         diff_bin7 = bin_7 - BL, 
         diff_bin8 = bin_8 - BL, 
         diff_bin9 = bin_9 - BL, 
         diff_bin10 = bin_10 - BL) %>%
  select(-BL, - starts_with("bin"))

# going to make it long now 
face_brow_z_diff_long <- face_brow_z_diff %>%
  pivot_longer(names_to = "bin", 
               values_to = "Zdiff", 
               diff_bin1:diff_bin10)
  
```

#### checking on missingness
```{r}
library(naniar)

vis_miss(face_brow_z_diff_long)
```



#### plotting - face brow z-diff 
```{r}
# make conditio a factor 
face_brow_z_diff_long$condition <- as.factor(face_brow_z_diff_long$condition)

face_brow_z_diff_long$bin <- fct_relevel(face_brow_z_diff_long$bin, c("diff_bin1", "diff_bin2", "diff_bin3", "diff_bin4", "diff_bin5", "diff_bin6", "diff_bin7", "diff_bin8", "diff_bin9", "diff_bin10"))

face_brow_z_diff_long %>%
  group_by(condition, bin) %>%
  summarise(meanBROWdiff = mean(Zdiff, na.rm = TRUE)) %>%
  ggplot(aes(x = bin, y = meanBROWdiff, colour = condition, group = condition)) +
  geom_point() + 
  geom_line() +
  labs(title = "brow activity difference from baseline in response to faces") +
  scale_y_continuous(limits = c(-0.2, 0.2))

```

### faces - cheek
```{r}
# add a muscle column to make things easier to join back together later
face_cheek_z <- faces_data_zscore %>%
  mutate(muscle = "cheek") %>%
  select(pp_no, condition, bin, trial, muscle, Zcheek) 

# Make the bin column wide and rename bin_0 as BL (i.e. baseline)
face_cheek_z_wide <- face_cheek_z %>%
  pivot_wider(names_from = bin, 
              values_from = Zcheek) %>%
  rename(BL = bin_0)

# mutate() diff scores - Use wide columns to calcuate the difference between each bin column and BL, creating a new set of columns starting with “diff”, drop BL column and all columns starting with bin (i.e. raw z scores).

# note - This df contains for each bin the difference between stimulus and basline, so POSITIVE difference scores = greater activity during STIM than BL and NEGATIVE difference scores = greater activity during BL than STIM

face_cheek_z_diff <- face_cheek_z_wide %>% 
  mutate(diff_bin1 = bin_1 - BL, 
         diff_bin2 = bin_2 - BL,
         diff_bin3 = bin_3- BL, 
         diff_bin4 = bin_4 - BL,
         diff_bin5 = bin_5 - BL, 
         diff_bin6 = bin_6 - BL,
         diff_bin7 = bin_7 - BL, 
         diff_bin8 = bin_8 - BL, 
         diff_bin9 = bin_9 - BL, 
         diff_bin10 = bin_10 - BL) %>%
  select(-BL, - starts_with("bin"))

# going to make it long now 
face_cheek_z_diff_long <- face_cheek_z_diff %>%
  pivot_longer(names_to = "bin", 
               values_to = "Zdiff", 
               diff_bin1:diff_bin10)
  
```

#### plotting - face cheek z-diff 
```{r}
# make condition a factor 
face_cheek_z_diff_long$condition <- as.factor(face_cheek_z_diff_long$condition)

face_cheek_z_diff_long$bin <- fct_relevel(face_cheek_z_diff_long$bin, c("diff_bin1", "diff_bin2", "diff_bin3", "diff_bin4", "diff_bin5", "diff_bin6", "diff_bin7", "diff_bin8", "diff_bin9", "diff_bin10"))

face_cheek_z_diff_long %>%
  group_by(condition, bin) %>%
  summarise(meanCHEEKdiff = mean(Zdiff, na.rm = TRUE)) %>%
  ggplot(aes(x = bin, y = meanCHEEKdiff, colour = condition, group = condition)) +
  geom_point() + 
  geom_line() +
  labs(title = "cheek activity difference from baseline in response to faces") +
  scale_y_continuous(limits = c(-0.2, 0.3))
```

## non-facial stimuli 

moving onto non-facial stimuli!

### non-face - brow

```{r}
# add a muscle column to make things easier to join back together later
non_brow_z <- non_data_zscore %>%
  mutate(muscle = "brow") %>%
  select(pp_no, condition, bin, trial, muscle, Zbrow) 

# Make the bin column wide and rename bin_0 as BL (i.e. baseline)
non_brow_z_wide <- non_brow_z %>%
  pivot_wider(names_from = bin, 
              values_from = Zbrow) %>%
  rename(BL = bin_0)

# mutate() diff scores - Use wide columns to calcuate the difference between each bin column and BL, creating a new set of columns starting with “diff”, drop BL column and all columns starting with bin (i.e. raw z scores).

# note - This df contains for each bin the difference between stimulus and basline, so POSITIVE difference scores = greater activity during STIM than BL and NEGATIVE difference scores = greater activity during BL than STIM

non_brow_z_diff <- non_brow_z_wide %>% 
  mutate(diff_bin1 = bin_1 - BL, 
         diff_bin2 = bin_2 - BL,
         diff_bin3 = bin_3 - BL, 
         diff_bin4 = bin_4 - BL,
         diff_bin5 = bin_5 - BL, 
         diff_bin6 = bin_6 - BL,
         diff_bin7 = bin_7 - BL, 
         diff_bin8 = bin_8 - BL, 
         diff_bin9 = bin_9 - BL, 
         diff_bin10 = bin_10 - BL) %>%
  select(-BL, - starts_with("bin"))

# going to make it long now 
non_brow_z_diff_long <- non_brow_z_diff %>%
  pivot_longer(names_to = "bin", 
               values_to = "Zdiff", 
               diff_bin1:diff_bin10)
  
```

#### plotting - non brow z-diff 
```{r}
# make conditio a factor 
non_brow_z_diff_long$condition <- as.factor(non_brow_z_diff_long$condition)

non_brow_z_diff_long$bin <- fct_relevel(non_brow_z_diff_long$bin, c("diff_bin1", "diff_bin2", "diff_bin3", "diff_bin4", "diff_bin5", "diff_bin6", "diff_bin7", "diff_bin8", "diff_bin9", "diff_bin10"))

non_brow_z_diff_long %>%
  group_by(condition, bin) %>%
  summarise(meanBROWdiff = mean(Zdiff, na.rm = TRUE)) %>%
  ggplot(aes(x = bin, y = meanBROWdiff, colour = condition, group = condition)) +
  geom_point() + 
  geom_line() +
  labs(title = "brow activity difference from baseline in response to non-facial stimuli") +
  scale_y_continuous(limits = c(-0.2, 0.2))
```

### non-face - cheek 

```{r}
# add a muscle column to make things easier to join back together later
non_cheek_z <- non_data_zscore %>%
  mutate(muscle = "cheek") %>%
  select(pp_no, condition, bin, trial, muscle, Zcheek) 

# Make the bin column wide and rename bin_0 as BL (i.e. baseline)
non_cheek_z_wide <- non_cheek_z %>%
  pivot_wider(names_from = bin, 
              values_from = Zcheek) %>%
  rename(BL = bin_0)

# mutate() diff scores - Use wide columns to calcuate the difference between each bin column and BL, creating a new set of columns starting with “diff”, drop BL column and all columns starting with bin (i.e. raw z scores).

# note - This df contains for each bin the difference between stimulus and basline, so POSITIVE difference scores = greater activity during STIM than BL and NEGATIVE difference scores = greater activity during BL than STIM

non_cheek_z_diff <- non_cheek_z_wide %>% 
  mutate(diff_bin1 = bin_1 - BL, 
         diff_bin2 = bin_2 - BL,
         diff_bin3 = bin_3 - BL, 
         diff_bin4 = bin_4 - BL,
         diff_bin5 = bin_5 - BL, 
         diff_bin6 = bin_6 - BL,
         diff_bin7 = bin_7 - BL, 
         diff_bin8 = bin_8 - BL, 
         diff_bin9 = bin_9 - BL, 
         diff_bin10 = bin_10 - BL) %>%
  select(-BL, - starts_with("bin"))

# going to make it long now 
non_cheek_z_diff_long <- non_cheek_z_diff %>%
  pivot_longer(names_to = "bin", 
               values_to = "Zdiff", 
               diff_bin1:diff_bin10)
  
```

#### plotting - non cheek z-diff 
```{r}
# make condition a factor 
non_cheek_z_diff_long$condition <- as.factor(non_cheek_z_diff_long$condition)

non_cheek_z_diff_long$bin <- fct_relevel(non_cheek_z_diff_long$bin, c("diff_bin1", "diff_bin2", "diff_bin3", "diff_bin4", "diff_bin5", "diff_bin6", "diff_bin7", "diff_bin8", "diff_bin9", "diff_bin10"))

non_cheek_z_diff_long %>%
  group_by(condition, bin) %>%
  summarise(meanCHEEKdiff = mean(Zdiff, na.rm = TRUE)) %>%
  ggplot(aes(x = bin, y = meanCHEEKdiff, colour = condition, group = condition)) +
  geom_point() + 
  geom_line() +
  labs(title = "cheek activity difference from baseline in response to non-facial stimuli")
```


## going to check raw BL scores - grouped by condition and muscle for face and non-face data
```{r}
# use face data 
face_BL_stat <- faces_data %>%
  filter(bin == "bin_0") %>%
  group_by(condition, muscle) %>%
  summarise(meanRMS = mean(rms)) %>%
  arrange(muscle)

# non-face data 
non_BL_stat <- non_data %>%
  filter(bin == "bin_0") %>%
  group_by(condition, muscle) %>%
  summarise(meanRMS = mean(rms)) %>%
  arrange(muscle)
```


# combine cheek and brow data for face and non-face, write_csv. 

```{r}
face_zdiff <- bind_rows(face_brow_z_diff_long, face_cheek_z_diff_long) %>%
  arrange(pp_no, condition, trial, muscle, bin) %>%
  write.csv(here("R_analysing", "data", "8_face_zdiff.csv"))

non_zdiff <- bind_rows(non_brow_z_diff_long, non_cheek_z_diff_long) %>%
  arrange(pp_no, condition, trial, muscle, bin) %>%
  write.csv(here("R_analysing", "data", "8_non_zdiff.csv"))
```
