==> 1_face_non/R_analysing/scripts/11b_lmm_face_analyses.Rmd <==
---
title: "11.2 lmm_face_analyses"
author: "Sahar Fatima"
date: "26/08/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


# Objective:
This script contains LMMs for cheek and brow responding to facial stimuli. 

I will begin with the maximal global model (model that  based on theoretical research - i.e.,  contains variables related to the matter at hand) for cheek and brow responding separately, and if it doesn't converge will simplify. 

According to Harrison et al. (2018), "the global model should be carefully constructed using the researchers’ knowledge and understanding of the system such that only predictors likely to be pertinent to the problem at hand are included, rather than including all the predictors the researcher has collected and/or has available."


## load packages 

```{r message=FALSE, warning=FALSE}
library(here)
library(tidyverse)
library(lme4)
library(lmerTest)
library(ggeffects)
library(broom.mixed)
library(pixiedust)
library(beepr)
library(effsize)
library(lsmeans)
```

## read in data 

```{r message=FALSE, warning=FALSE}
face <- read_csv(here("R_analysing", "data", "9_face_zdiff_screened.csv")) %>%
  mutate(condition = recode(condition, "429" = "angry","673" = "happy", .default = "NA"))

options(scipen = 999)
```

## fix data types 

```{r}
face$condition <- as.factor(face$condition)

face <- face %>% 
  mutate_if(is.character,as.factor) 

levels(face$bin)

face$bin <- fct_relevel(face$bin, c("diff_bin1", "diff_bin2", "diff_bin3", "diff_bin4", "diff_bin5", "diff_bin6", "diff_bin7", "diff_bin8", "diff_bin9", "diff_bin10"))

```

## make cheek and brow data frames
```{r}
#face_cheek 
face_cheek <- face %>%
  filter(muscle == "cheek") %>%
  arrange(pp_no, condition, trial, bin)

#face_brow
face_brow <- face %>%
  filter(muscle == "brow") %>%
  arrange(pp_no, condition, trial, bin)
```


# CHEEK 

### Cheek MODEL 1 (including all main effects and interactions as fixed effects and random slopes)
This model predicts Zdiff from fixed effects of condition (happy, angry) and bin (1-10), and their interactions (condition x bin).

It includes random intercepts for participant (accounting for the potential of some kids to just have more active faces than others) and random slopes for condition, bin and their interaction (condition*bin). 

```{r}
# did not converge 

# lm_face_cheek_1 <- lmer(Zdiff ~ condition + bin + condition*bin + (1 + condition + bin + condition*bin|pp_no), data = face_cheek, REML = FALSE) 
```

#### Model 1 take home 
Cheek Model 1 did not converge - so going to simplify by removing random slope for the interaction. 

### Cheek MODEL 2 (random slopes for condition and bin)
It includes random intercepts for participant (accounting for the potential of some kids to just have more active faces than others) and random slopes for condition and bin. 


```{r}
# converges! took some time (15 mins)

# lm_face_cheek_2 <- lmer(Zdiff ~ condition + bin + condition*bin + (1 + condition + bin|pp_no), data = face_cheek, REML = FALSE)
```

##### check if normality assumptions are met 
``` {r}
# plot(lm_face_cheek_2)

# qqnorm(resid(lm_face_cheek_2))
# qqline(resid(lm_face_cheek_2)) # spoiler alert they aren't

```

#### transform to correct normality
Usually we would log transform but the Zdiff includes negative values. Alternative is to use a log modulus transform. This transformation log transforms the absolute value (without the -ve) and then puts the sign back on). 
``` {r}
face_cheek <- face_cheek %>%
  mutate(log_modulus = sign(Zdiff) * log(1+abs(Zdiff)))

# run model again

# lm_face_cheek_2 <- lmer(log_modulus ~ condition + bin + condition*bin + (1 + condition + bin|pp_no), data = face_cheek, REML = FALSE)
```

#### test assumptions again 
Residuals are more evenly distributed than before; qq plot is better as well.

```{r}
# plot(lm_face_cheek_2)

# qqnorm(resid(lm_face_cheek_2))
# qqline(resid(lm_face_cheek_2))

```

### get anova() and summary
```{r}
# use anova to estimate effects
# aov_face_cheek_2 <- anova(lm_face_cheek_2) %>%
  # rownames_to_column() %>%
  # rename(term = rowname)

# use summary to get coefficients
# summary(lm_face_cheek_2)
```


#### model 2 take home 
Model 2 converges - neither the main effects nor the interaction is sig. 

Going to compare with simpler model just to compare fit (with slope for condition only)

### Cheek MODEL 3 (random slopes for condition)
It includes random intercepts for participant (accounting for the potential of some kids to just have more active faces than others) and random slopes for condition.

```{r}
# converges! 
lm_face_cheek_3 <- lmer(log_modulus ~ condition + bin + condition*bin + (1 + condition|pp_no), data = face_cheek, REML = FALSE)
```

#### compare fit
compare fit with model 2 (random slopes for condition and bin)
```{r}
# AIC(lm_face_cheek_2, lm_face_cheek_3)
```

#### likelihood ratio test
Use anova to test whether the difference in fit is significant. 
```{r}
# anova(lm_face_cheek_2, lm_face_cheek_3)
```

#### get anova()/ summary
```{r}
# use anova to estimate effects
aov_face_cheek_3 <- anova(lm_face_cheek_3) %>%
  rownames_to_column() %>%
  rename(term = rowname)

anova_cheek <- anova(lm_face_cheek_3)

# use summary to get coefficients
summary(lm_face_cheek_3)

confint.merMod(lm_face_cheek_3, level = 0.95)

```


#### model 3 take home 
so Model 3 converges - neither the main effects nor the interaction is sig. 

In summary table, interaction b/ween conditionhappy:bindiff_bin10 and b/ween conditionhappy:bindiff_bin9 is close to significance(?) Which [I think] means that there's a trend for baseline cheek responding to be greater in response to happy faces relative to angry faces in bin_9 and bin_10 which is 900 - 1000ms into stimulus presentation. 

Going to compare with simpler model just to compare fit. 

### Cheek MODEL 4 (intercepts only)

```{r}
lm_face_cheek_4 <- lmer(log_modulus ~ condition + bin + condition*bin + (1|pp_no), data = face_cheek, REML = FALSE)
```

#### compare fit
```{r}
AIC(lm_face_cheek_3, lm_face_cheek_4)
```
#### model 4 take home 

Model 3 is better fit than Model 4. 

## Cheek TAKE HOME
Model 3 with fixed effects of condition, bin and their interaction and slopes for condition has best fit. 

Model 3 - neither the main effect nor the interaction is significant. 


# BROW

### Brow MODEL 1 (including all main effects and interactions as fixed effects and random slopes)
This model predicts Zdiff from fixed effects of condition (happy, angry) and bin (1-10), and their interactions (condition x bin).

It includes random intercepts for participant (accounting for the potential of some kids to just have more active faces than others) and random slopes for condition, bin and their interaction (condition*bin). 

```{r}
# did not converge 

# lm_face_brow_1 <- lmer(Zdiff ~ condition + bin + condition*bin + (1 + condition + bin + condition*bin|pp_no), data = face_brow, REML = FALSE) 
```

#### Model 1 take home 
Brow Model 1 did not converge - so going to simplify by removing random slope for the interaction. 

### Brow MODEL 2 (random slopes for condition and bin)
It includes random intercepts for participant (accounting for the potential of some kids to just have more active faces than others) and random slopes for condition and bin. 


```{r}
# converges!

# lm_face_brow_2 <- lmer(Zdiff ~ condition + bin + condition*bin + (1 + condition + bin|pp_no), data = face_brow, REML = FALSE)
```

##### check if normality assumptions are met 
``` {r}
# plot(lm_face_brow_2)

# qqnorm(resid(lm_face_brow_2))
# qqline(resid(lm_face_brow_2)) # spoiler alert they aren't

```

#### transform to correct normality
Usually we would log transform but the Zdiff includes negative values. Alternative is to use a log modulus transform. This transformation log transforms the absolute value (without the -ve) and then puts the sign back on). 
``` {r}
face_brow <- face_brow %>%
  mutate(log_modulus = sign(Zdiff) * log(1+abs(Zdiff)))

# run model again - fails to converge

# lm_face_brow_2 <- lmer(log_modulus ~ condition + bin + condition*bin + (1 + condition + bin|pp_no), data = face_brow, REML = FALSE)
```


#### test assumptions again 
Residuals are more evenly distributed than before; qq plot is better as well.

```{r}
# plot(lm_face_brow_2)

# qqnorm(resid(lm_face_brow_2))
# qqline(resid(lm_face_brow_2))

```



#### model 2 take home 
Model 2 converged before log transforming z-scores to make distribution novel.

Going to compare with simpler model just to compare fit (with slope for condition only)

### Brow MODEL 3 (slopes for condition)

```{r}
lm_face_brow_3 <- lmer(log_modulus ~ condition + bin + condition*bin + (1 + condition|pp_no), data = face_brow, REML = FALSE)
```


### get anova()/ summary
```{r}
aov_face_brow_3 <- anova(lm_face_brow_3) %>%
  rownames_to_column() %>%
  rename(term = rowname)

anova(lm_face_brow_3)
summary(lm_face_brow_3)

confint.merMod(lm_face_brow_3, level = 0.95)

```

#### model 3 take home
Brow model 3 converges - main effect for bin is significant. 

from summary conditionhappy:bindiff_bin10 is significant - suggesting that the difference from baseline brow responding is greater in response to angry faces relative to happy faces at bin_10 (direction determined from plot in 8. calculating_zscores script). 

Going to compare with simpler model just to compare fit. 

### Brow MODEL 4 (intercepts only)

```{r}
lm_face_brow_4 <- lmer(log_modulus ~ condition + bin + condition*bin + (1|pp_no), data = face_brow, REML = FALSE)
```

#### compare fit
```{r}
AIC(lm_face_brow_3, lm_face_brow_4)
```
#### model 4 take home 
Model 3 is better fit than Model 4. 

## Brow TAKE HOME
Model 3 with fixed effects of condition, bin and their interaction, and slopes for condition has best fit. 

Model 3 main effect for bin is significant (from anova output)

from summary conditionhappy:bindiff_bin10 is significant - suggesting that the difference from baseline brow responding is greater in response to angry faces relative to happy faces at bin_10 (direction determined from plot in 8. calculating_zscores script).
